Move check to separate file
