+++ packages/checkbox/src/Checkbox.tsx
@@ -7,10 +7,9 @@ import { useIdAllocator } from '@leafygreen-ui/hooks';
 import { css, cx } from '@leafygreen-ui/emotion';
 import { palette, uiColors } from '@leafygreen-ui/palette';
 import { LegacyCheck } from './LegacyCheck';
-import SvgCheck from './SvgCheck';
-import SvgIndeterminate from './SvgIndeterminate';
-import { Transition, TransitionStatus } from 'react-transition-group';
 import { fontFamilies } from '@leafygreen-ui/tokens';
+import { Check } from './Check';
+import { checkAnimationDuration, checkBoxSize } from './constants';
 
 const checkboxWrapper = createDataProp('checkbox-wrapper');
 const checkboxInput = createDataProp('checkbox-input');
@@ -21,14 +20,6 @@ const Mode = {
 } as const;
 
 type Mode = typeof Mode[keyof typeof Mode];
-
-const checkBoxSize = 14;
-const checkAnimationDuration = 100;
-const hypotenusePct = 100 * Math.sqrt(2); // relative distance from corner to corner
-const insetPct = 100 - hypotenusePct;
-const flourishScale = 2.25;
-const flourishTransitionScale = 4;
-const flourishTransitionDelay = -flourishTransitionScale / flourishScale;
 
 const containerStyle = css`
   --lg-checkbox-base-duration: 0ms;
@@ -103,123 +94,6 @@ const inputFocusStylesDarkMode = css`
     box-shadow: unset;
   }
 `;
-
-const checkWrapperBaseStyle = css`
-  --lg-checkbox-border-color: ${palette.gray.dark2};
-  grid-area: check;
-  position: relative;
-  display: flex;
-  z-index: 1;
-  height: ${checkBoxSize}px;
-  width: ${checkBoxSize}px;
-  align-items: center;
-  justify-content: center;
-  border-radius: 3px;
-  border: 2px solid var(--lg-checkbox-border-color);
-  overflow: hidden;
-  background-color: transparent;
-  transition: box-shadow 100ms ease-in-out, background-color 0ms linear,
-    border-color 0ms linear;
-
-  // delay border-color out
-  transition-delay: 0ms, 0ms, var(--lg-checkbox-base-duration);
-
-  /**
-   * The animated background circle
-   */
-  &:before {
-    content: '';
-    position: absolute;
-    inset: ${insetPct}%;
-    z-index: 1;
-    border-radius: 100%;
-    background-color: ${palette.blue.base};
-    transform: scale(0);
-    transform-origin: center center;
-    transition: transform var(--lg-checkbox-base-duration) ease-in-out;
-    transition-delay: calc(var(--lg-checkbox-base-duration) / 2);
-  }
-`;
-
-const checkWrapperCheckedStyle = css`
-  --lg-checkbox-border-color: ${palette.blue.base};
-  background-color: ${palette.blue.base};
-  // Delay background transition in
-  transition-delay: 0ms, var(--lg-checkbox-base-duration), 0ms;
-
-  &:before {
-    transform: scale(1);
-    // No delay on enter
-    transition-delay: 0ms;
-  }
-`;
-
-const checkWrapperDisabledStyle = css`
-  --lg-checkbox-border-color: ${palette.gray.light2};
-  background-color: ${palette.gray.light3};
-  box-shadow: unset;
-
-  &:before {
-    background-color: ${palette.gray.light3};
-  }
-`;
-
-const checkWrapperCheckedDisabledStyle = css`
-  background-color: var(--lg-checkbox-border-color);
-
-  &:before {
-    background-color: var(--lg-checkbox-border-color);
-  }
-`;
-
-const flourishStyles = css`
-  grid-area: check;
-  height: ${checkBoxSize}px;
-  width: ${checkBoxSize}px;
-  z-index: 0;
-  border-radius: 100%;
-  background-color: ${palette.blue.dark1};
-  opacity: 1;
-  transform: scale(0);
-  transform-origin: center center;
-  transition: 0ms ease-in-out;
-  transition-property: transform, opacity;
-`;
-
-const flourishStylesChecked = css`
-  // only animate flourish on enter
-  transition-duration: calc(
-    ${flourishTransitionScale} * var(--lg-checkbox-base-duration)
-  );
-  transition-delay: calc(
-    ${flourishTransitionDelay} * var(--lg-checkbox-base-duration)
-  );
-  transform: scale(${flourishScale});
-  opacity: 0;
-`;
-
-const checkIconStyles = css`
-  z-index: 1;
-  transform-origin: center;
-  transition: transform var(--lg-checkbox-base-duration) ease-in-out;
-`;
-
-const checkInStyles = css`
-  transform: scale(1) rotate(0);
-  transition-delay: calc(var(--lg-checkbox-base-duration) / 8);
-`;
-
-const checkOutStyles = css`
-  transform: scale(0) rotate(-45deg);
-`;
-
-const checkIconTransitionStyles: Record<TransitionStatus, string> = {
-  entering: checkOutStyles,
-  entered: checkInStyles,
-  exiting: checkOutStyles,
-  exited: checkOutStyles,
-  unmounted: checkOutStyles,
-};
 
 const baseLabelStyle = css`
   grid-area: label;
@@ -310,18 +184,13 @@ function Checkbox({
     }
   };
 
-  const CheckIcon = indeterminateProp ? SvgIndeterminate : SvgCheck;
-  const showCheckIcon = indeterminateProp || isChecked;
-  const checkIconColor = disabled ? palette.gray.light3 : palette.white;
-  const shouldAnimate = animate;
-
   return (
     <Label
       className={cx(
         containerStyle,
         {
           [disabledContainerStyle]: disabled,
-          [enableAnimationStyles]: shouldAnimate,
+          [enableAnimationStyles]: animate,
           // TODO: Refresh - remove darkMode logic
           [css`
             &:hover
@@ -361,44 +230,19 @@ function Checkbox({
       {darkMode ? (
         <LegacyCheck
           isChecked={isChecked}
-          indeterminateProp={indeterminateProp}
+          indeterminate={indeterminateProp}
           disabled={disabled}
           animate={animate}
-          checkboxWrapper={checkboxWrapper}
+          selector={checkboxWrapper}
         />
       ) : (
-        <>
-          <div
-            {...checkboxWrapper.prop}
-            className={cx(checkWrapperBaseStyle, {
-              [checkWrapperCheckedStyle]: showCheckIcon,
-              [checkWrapperDisabledStyle]: disabled,
-              [checkWrapperCheckedDisabledStyle]: disabled && showCheckIcon,
-            })}
-          >
-            <Transition
-              in={showCheckIcon}
-              timeout={shouldAnimate ? checkAnimationDuration : 0}
-            >
-              {state => (
-                <CheckIcon
-                  stroke={checkIconColor}
-                  className={cx(
-                    checkIconStyles,
-                    checkIconTransitionStyles[state],
-                  )}
-                />
-              )}
-            </Transition>
-          </div>
-          {!usingKeyboard && (
-            <div
-              className={cx(flourishStyles, {
-                [flourishStylesChecked]: isChecked,
-              })}
-            />
-          )}
-        </>
+        <Check
+          isChecked={isChecked}
+          indeterminate={indeterminateProp}
+          disabled={disabled}
+          animate={animate}
+          selector={checkboxWrapper}
+        />
       )}
 
       {label && (
