+++ packages/checkbox/src/constants.ts
@@ -1,0 +1,7 @@
+export const checkBoxSize = 14;
+export const checkAnimationDuration = 100;
+export const hypotenusePct = 100 * Math.sqrt(2); // relative distance from corner to corner
+export const insetPct = 100 - hypotenusePct;
+export const flourishScale = 2.25;
+export const flourishTransitionScale = 4;
+export const flourishTransitionDelay = -flourishTransitionScale / flourishScale;
