+++ packages/checkbox/src/types.ts
@@ -1,0 +1,9 @@
+import { createDataProp } from '@leafygreen-ui/lib';
+
+export interface CheckProps {
+  isChecked: boolean;
+  disabled: boolean;
+  animate: boolean;
+  selector: ReturnType<typeof createDataProp>;
+  indeterminate?: boolean;
+}
